move files from debug to algorithms
