+++ src/main/scala/whu/edu/cn/algorithms/SpatialStats/GWModels/readme
