+++ src/main/scala/whu/edu/cn/algorithms/SpatialStats/SpatialHeterogeneity/readme
