+++ src/main/scala/whu/edu/cn/algorithms/SpatialStats/SpatialRegression/PoissonRegression.scala
@@ -1,4 +1,4 @@
-package whu.edu.cn.debug.SpatialStats.SpatialRegression
+package whu.edu.cn.algorithms.SpatialStats.SpatialRegression
 
 import breeze.linalg.DenseVector
