+++ src/main/scala/whu/edu/cn/algorithms/SpatialStats/SpatialRegression/SpatialDurbinModel.scala
@@ -1,10 +1,10 @@
-package whu.edu.cn.debug.SpatialStats.SpatialRegression
+package whu.edu.cn.algorithms.SpatialStats.SpatialRegression
 
 import breeze.linalg.{DenseMatrix, DenseVector, eig, inv, qr, sum}
 import breeze.numerics.sqrt
 import scala.math._
 
-import whu.edu.cn.debug.SpatialStats.Utils.Optimize._
+import whu.edu.cn.algorithms.SpatialStats.Utils.Optimize._
 
 /**
  * 空间杜宾模型，同时考虑自变量误差项λ与因变量滞后项ρ。
