+++ src/main/scala/whu/edu/cn/algorithms/SpatialStats/SpatialRegression/SpatialLagModel.scala
@@ -1,9 +1,9 @@
-package whu.edu.cn.debug.SpatialStats.SpatialRegression
+package whu.edu.cn.algorithms.SpatialStats.SpatialRegression
 
 import breeze.linalg._
 import scala.math._
 
-import whu.edu.cn.debug.SpatialStats.Utils.Optimize._
+import whu.edu.cn.algorithms.SpatialStats.Utils.Optimize._
 
 /**
  * 空间滞后模型，考虑因变量滞后项ρ。
