+++ src/main/scala/whu/edu/cn/algorithms/SpatialStats/SpatialRegression/readme
